requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-WTF
Flask-Cors
Flask-Caching # Используется Flask-Caching вместо Flask-Cache
Flask-Limiter
Flask-SocketIO
Flask-Redis
gunicorn
eventlet
pydantic
python-dotenv
prometheus-flask-exporter
numpy
scipy
fastdtw
pydantic
python-json-logger
pytest
pytest-cov
pytest-flask
pytest-mock
Unidecode
Flask-Login
Werkzeug
email-validator
